implemented PriorityType should be unsigned integral

+++ tests/containertestlib.cc
@@ -12,7 +12,7 @@
 
 TEST_CASE( "Basic Operations for Max Heap", "[main]" )
 {
-    auto b = BucketQueue<std::string, int>();
+    auto b = BucketQueue<std::string, unsigned short>();
 
     std::vector<std::string> elements {"low", "medium low",
         "medium", "high", "highest"};
@@ -21,7 +21,7 @@ TEST_CASE( "Basic Operations for Max Heap", "[main]" )
 
     auto itr = elements.crbegin();
 
-    int priority = 1;
+    unsigned short priority = 1;
     for (auto e : elements)
     {
         b.push(e, priority);
@@ -74,7 +74,7 @@ TEST_CASE( "Basic Operations for Max Heap", "[main]" )
 
 TEST_CASE( "Basic Operations for Min Heap", "[main]" )
 {
-    auto b = BucketQueue<std::string, int, std::greater<int>()>();
+    auto b = BucketQueue<std::string, unsigned short, std::greater<unsigned short>()>();
 
     std::vector<std::string> elements {"highest", "high",
         "medium", "medium low", "low"};
@@ -83,7 +83,7 @@ TEST_CASE( "Basic Operations for Min Heap", "[main]" )
 
     auto itr = elements.cbegin();
 
-    int priority = 1;
+    unsigned short priority = 1;
     for (auto e : elements)
     {
         b.push(e, priority);
@@ -136,7 +136,7 @@ TEST_CASE( "Basic Operations for Min Heap", "[main]" )
 
 TEST_CASE( "Same Priority", "[main]" )
 {
-    auto b = BucketQueue<std::string, int, std::greater<int>()>();
+    auto b = BucketQueue<std::string, unsigned short, std::greater<unsigned short>()>();
 
     std::vector<std::string> samePriorityElements {"medium1", "medium2"};
 
@@ -161,7 +161,7 @@ TEST_CASE( "Same Priority", "[main]" )
 TEST_CASE( "Basic Operations for Min Heap Node", "[main]" )
 {
 
-    auto b = BucketQueue<Node, unsigned int, std::greater<Node>()>();
+    auto b = BucketQueue<Node, unsigned unsigned short, std::greater<Node>()>();
     constexpr std::size_t N = 3;
 
     Node startNode({8, 4, 2, 1, constants::EMPTY, 5, 6, 7, 3});
